+++ examples/tutorial_pose/2_extract_pose_or_heatmat_from_image.cpp
@@ -51,7 +51,7 @@ DEFINE_int32(scale_number,              1,              "Number of scales to ave
 // OpenPose Rendering
 DEFINE_int32(part_to_show,              19,             "Prediction channel to visualize (default: 0). 0 for all the body parts, 1-18 for each body"
                                                         " part heat map, 19 for the background heat map, 20 for all the body part heat maps"
-                                                        " together, 21 for all the PAFs, 22-40 for each body part pair PAF");
+                                                        " together, 21 for all the PAFs, 22-40 for each body part pair PAF.");
 DEFINE_bool(disable_blending,           false,          "If enabled, it will render the results (keypoint skeletons or heatmaps) on a black"
                                                         " background, instead of being rendered into the original image. Related: `part_to_show`,"
                                                         " `alpha_pose`, and `alpha_pose`.");