+++ examples/tutorial_wrapper/2_user_synchronous.cpp
@@ -56,6 +56,12 @@ DEFINE_int32(keypoint_scale,            0,              "Scaling of the (x,y) co
                                                         " size (set with `net_resolution`), `2` to scale it to the final output size (set with"
                                                         " `resolution`), `3` to scale it in the range [0,1], and 4 for range [-1,1]. Non related"
                                                         " with `scale_number` and `scale_gap`.");
+DEFINE_int32(number_people_max,         -1,             "This parameter will limit the maximum number of people detected, by keeping the people with"
+                                                        " top scores. The score is based in person area over the image, body part score, as well as"
+                                                        " joint score (between each pair of connected body parts). Useful if you know the exact"
+                                                        " number of people in the scene, so it can remove false positives (if all the people have"
+                                                        " been detected. However, it might also include false negatives by removing very small or"
+                                                        " highly occluded people. -1 will keep them all.");
 // OpenPose Body Pose
 DEFINE_bool(body_disable,               false,          "Disable body keypoint detection. Option only possible for faster (but less accurate) face"
                                                         " keypoint detection.");
@@ -106,18 +112,18 @@ DEFINE_bool(hand,                       false,          "Enables hand keypoint d
 DEFINE_string(hand_net_resolution,      "368x368",      "Multiples of 16 and squared. Analogous to `net_resolution` but applied to the hand keypoint"
                                                         " detector.");
 DEFINE_int32(hand_scale_number,         1,              "Analogous to `scale_number` but applied to the hand keypoint detector. Our best results"
-                                                        " were found with `hand_scale_number` = 6 and `hand_scale_range` = 0.4");
+                                                        " were found with `hand_scale_number` = 6 and `hand_scale_range` = 0.4.");
 DEFINE_double(hand_scale_range,         0.4,            "Analogous purpose than `scale_gap` but applied to the hand keypoint detector. Total range"
                                                         " between smallest and biggest scale. The scales will be centered in ratio 1. E.g. if"
                                                         " scaleRange = 0.4 and scalesNumber = 2, then there will be 2 scales, 0.8 and 1.2.");
 DEFINE_bool(hand_tracking,              false,          "Adding hand tracking might improve hand keypoints detection for webcam (if the frame rate"
                                                         " is high enough, i.e. >7 FPS per GPU) and video. This is not person ID tracking, it"
                                                         " simply looks for hands in positions at which hands were located in previous frames, but"
-                                                        " it does not guarantee the same person ID among frames");
+                                                        " it does not guarantee the same person ID among frames.");
 // OpenPose Rendering
 DEFINE_int32(part_to_show,              0,              "Prediction channel to visualize (default: 0). 0 for all the body parts, 1-18 for each body"
                                                         " part heat map, 19 for the background heat map, 20 for all the body part heat maps"
-                                                        " together, 21 for all the PAFs, 22-40 for each body part pair PAF");
+                                                        " together, 21 for all the PAFs, 22-40 for each body part pair PAF.");
 DEFINE_bool(disable_blending,           false,          "If enabled, it will render the results (keypoint skeletons or heatmaps) on a black"
                                                         " background, instead of being rendered into the original image. Related: `part_to_show`,"
                                                         " `alpha_pose`, and `alpha_pose`.");
@@ -424,7 +430,8 @@ int openPoseTutorialWrapper2()
                                                   poseModel, !FLAGS_disable_blending, (float)FLAGS_alpha_pose,
                                                   (float)FLAGS_alpha_heatmap, FLAGS_part_to_show, FLAGS_model_folder,
                                                   heatMapTypes, heatMapScale, FLAGS_part_candidates,
-                                                  (float)FLAGS_render_threshold, enableGoogleLogging};
+                                                  (float)FLAGS_render_threshold, FLAGS_number_people_max,
+                                                  enableGoogleLogging};
     // Face configuration (use op::WrapperStructFace{} to disable it)
     const op::WrapperStructFace wrapperStructFace{FLAGS_face, faceNetInputSize,
                                                   op::flagsToRenderMode(FLAGS_face_render, FLAGS_render_pose),