+++ include/openpose/3d/headers.hpp
@@ -1,0 +1,8 @@
+#ifndef OPENPOSE_3D_HEADERS_HPP
+#define OPENPOSE_3D_HEADERS_HPP
+
+// pose module
+#include <openpose/3d/poseTriangulation.hpp>
+#include <openpose/3d/wPoseTriangulation.hpp>
+
+#endif // OPENPOSE_3D_HEADERS_HPP