+++ include/openpose/core/headers.hpp
@@ -9,6 +9,7 @@
 #include <openpose/core/datum.hpp>
 #include <openpose/core/enumClasses.hpp>
 #include <openpose/core/gpuRenderer.hpp>
+#include <openpose/core/keepTopNPeople.hpp>
 #include <openpose/core/keypointScaler.hpp>
 #include <openpose/core/macros.hpp>
 #include <openpose/core/net.hpp>
@@ -24,6 +25,7 @@
 #include <openpose/core/scaleAndSizeExtractor.hpp>
 #include <openpose/core/wCvMatToOpInput.hpp>
 #include <openpose/core/wCvMatToOpOutput.hpp>
+#include <openpose/core/wKeepTopNPeople.hpp>
 #include <openpose/core/wKeypointScaler.hpp>
 #include <openpose/core/wOpOutputToCvMat.hpp>
 #include <openpose/core/wScaleAndSizeExtractor.hpp>