+++ include/openpose/experimental/3d/flirReader.hpp
@@ -2,7 +2,6 @@
 #define OPENPOSE_EXPERIMENTAL_3D_POINT_GREY_HPP
 
 #include <openpose/core/common.hpp>
-#include <openpose/experimental/3d/datum3D.hpp>
 #include <openpose/thread/workerProducer.hpp>
 
 namespace op
@@ -14,13 +13,13 @@ namespace op
      * 1. Set hardware trigger and the buffer to get the latest obtained frame.
      * 2. Read images from FLIR cameras.
      * 3. Turn them into std::vector<cv::Mat>.
-     * 4. Return the resulting images wrapped into a std::shared_ptr<std::vector<Datum3D>>.
+     * 4. Return the resulting images wrapped into a std::shared_ptr<std::vector<Datum>>.
      * The HW trigger + reading FLIR camera code is highly based on the Spinnaker SDK examples
      * `AcquisitionMultipleCamera` and specially `Trigger`
      * (located in `src/`). See them for more details about the cameras.
      * See `examples/tutorial_wrapper/` for more details about inhering the WorkerProducer class.
     */
-    class OP_API WFlirReader : public WorkerProducer<std::shared_ptr<std::vector<Datum3D>>>
+    class OP_API WFlirReader : public WorkerProducer<std::shared_ptr<std::vector<Datum>>>
     {
     public:
         WFlirReader();
@@ -29,7 +28,7 @@ namespace op
 
         void initializationOnThread();
 
-        std::shared_ptr<std::vector<Datum3D>> workProducer();
+        std::shared_ptr<std::vector<Datum>> workProducer();
 
     private:
         // PIMPL idiom