+++ include/openpose/experimental/3d/headers.hpp
@@ -3,12 +3,8 @@
 
 // Camera intrinsic and extrinsic parameters
 #include <openpose/experimental/3d/cameraParameters.hpp>
-// Datum3D
-#include <openpose/experimental/3d/datum3D.hpp>
 // Flir (Point-Grey) cameras
 #include <openpose/experimental/3d/flirReader.hpp>
-// 3D reconstruction
-#include <openpose/experimental/3d/reconstruction3D.hpp>
 // OpenGL Renderer
 #include <openpose/experimental/3d/renderer.hpp>