+++ include/openpose/experimental/3d/renderer.hpp
@@ -2,14 +2,13 @@
 #define OPENPOSE_EXPERIMENTAL_3D_RENDERER_HPP
 
 #include <openpose/core/common.hpp>
-#include <openpose/experimental/3d/datum3D.hpp>
 #include <openpose/pose/enumClasses.hpp>
 #include <openpose/thread/workerConsumer.hpp>
 
 namespace op
 {
     // This worker will do 3-D rendering
-    class OP_API WRender3D : public WorkerConsumer<std::shared_ptr<std::vector<Datum3D>>>
+    class OP_API WRender3D : public WorkerConsumer<std::shared_ptr<std::vector<Datum>>>
     {
     public:
         WRender3D(const PoseModel poseModel = PoseModel::COCO_18);
@@ -18,7 +17,7 @@ namespace op
 
         void initializationOnThread();
 
-        void workConsumer(const std::shared_ptr<std::vector<Datum3D>>& datumsPtr);
+        void workConsumer(const std::shared_ptr<std::vector<Datum>>& datumsPtr);
     };
 }