+++ src/openpose/3d/defineTemplates.cpp
@@ -1,0 +1,6 @@
+#include <openpose/3d/headers.hpp>
+
+namespace op
+{
+    DEFINE_TEMPLATE_DATUM(WPoseTriangulation);
+}