+++ src/openpose/core/defineTemplates.cpp
@@ -4,6 +4,7 @@ namespace op
 {
     DEFINE_TEMPLATE_DATUM(WCvMatToOpInput);
     DEFINE_TEMPLATE_DATUM(WCvMatToOpOutput);
+    DEFINE_TEMPLATE_DATUM(WKeepTopNPeople);
     DEFINE_TEMPLATE_DATUM(WKeypointScaler);
     DEFINE_TEMPLATE_DATUM(WOpOutputToCvMat);
     DEFINE_TEMPLATE_DATUM(WScaleAndSizeExtractor);